Add locale support in the mac bundle with updated scripts (translations are now supported on MacOSX builds)

+++ trunk/autogen.sh
@@ -49,8 +49,13 @@ datarootdir = @datarootdir@' po/Makefile.in.in > po/Makefile.in.in.tmp && mv -f 
         sed -e '/^datadir *=/a\
 datarootdir = @datarootdir@' intl/Makefile.in > intl/Makefile.in.tmp && mv -f intl/Makefile.in.tmp intl/Makefile.in
     fi
+UNAME=`uname`
+if [ x$UNAME == x"Darwin" ]; then
+    echo Not patching po/Makefile.in.in - sed is too old.
+else
     sed -e '/^clean:/a\
 	rm -f *.gmo' po/Makefile.in.in > po/Makefile.in.in.tmp && mv -f po/Makefile.in.in.tmp po/Makefile.in.in
+fi
 #    if [ -f Makefile -a -x config.status ]; then
 #        CONFIG_FILES=intl/Makefile CONFIG_HEADERS= /bin/sh ./config.status
 #    fi
